# docs/api/openapi-corvos-spec.yaml
openapi: 3.0.0
info:
  title: Corvos BJJ API
  description: |
    API para gerenciamento de academia de jiu-jitsu brasileiro.
    Endpoints para alunos, pagamentos, graduações e autenticação.
  version: 1.0.0
  contact:
    name: Corvos BJJ Team
    email: dev@corvos.com
  license:
    name: MIT
servers:
  - url: https://api.corvos.com
    description: Production
  - url: https://api-staging.corvos.com
    description: Staging
  - url: http://localhost:3001
    description: Development
...
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: professor@corvos.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePassword123!
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        tokens:
          type: object
          properties:
            accessToken:
              type: string
              description: JWT token (15 min expiry)
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            refreshToken:
              type: string
              description: Refresh token (7 day expiry, httpOnly cookie)
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: professor@corvos.com
        name:
          type: string
          example: João Silva
        role:
          type: string
          enum: [admin, professor]
          example: admin
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: 2026-01-15T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-01-15T10:30:00Z

    # ===== Student Schemas =====
...
    GradeInput:
      type: object
      required:
        - student_id
        - belt_color
        - promotion_date
      properties:
        student_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        belt_color:
          type: string
          enum: [branca, azul, roxa, marrom, preta]
          example: azul
        promotion_date:
          type: string
          format: date
          example: "2025-01-15"
        notes:
          type: string
          nullable: true
          example: "Excelente técnica, progresso consistente"
    Grade:
      allOf:
        - $ref: '#/components/schemas/GradeInput'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440003
            promoted_by:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
            created_at:
              type: string
              format: date-time
              example: 2025-01-15T14:00:00Z
            updated_at:
              type: string
              format: date-time
              example: 2025-01-15T14:00:00Z

    # ===== Paginated Response Schemas =====
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: Email is required
            details:
              type: object
              nullable: true
    UnauthorizedError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Invalid credentials or token expired
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token for API requests (15 min expiry)
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login with email and password
      description: Authenticate user and receive access + refresh tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abcdef123456; HttpOnly; Secure; SameSite=Strict; Max-Age=604800
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Logout current user
      description: Invalidate session and tokens
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current authenticated user
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Use refresh token to get new access token (no re-login needed)
      operationId: refreshToken
      responses:
        '200':
          description: New tokens generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tokens:
                    type: object
                    properties:
                      accessToken:
                        type: string
          headers:
            Set-Cookie:
              schema:
                type: string
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

  # ===== STUDENTS ENDPOINTS =====
...
  /grades:
    post:
      tags:
        - Grades
      summary: Record new promotion
      description: Register a new belt promotion for a student
      operationId: createGrade
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeInput'
      responses:
        '201':
          description: Promotion recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Grade'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /grades/{id}:
    put:
      tags:
        - Grades
      summary: Update grade record
      operationId: updateGrade
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeInput'
      responses:
        '200':
          description: Grade updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Grade'

    delete:
      tags:
        - Grades
      summary: Delete grade record
      operationId: deleteGrade
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Grade deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /students/{id}/grades:
    get:
      tags:
        - Grades
      summary: Get grade history for student
      description: Retrieve all belt promotions for a student
      operationId: getStudentGrades
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Grade history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Grade'

  # ===== REPORTS ENDPOINTS =====
...